Using class instead of function to generate datas
